refactor(Legend): update import paths to use '@/'
