+++ src/components/index.ts
@@ -1,2 +1,1 @@
-export * from "./legend";
 export * from "./sequencer";
