--- src/components/legend/components/mute/index.ts
@@ -1,1 +1,0 @@
-export * from "./MuteButton";
