--- src/components/legend/index.ts
@@ -1,1 +1,0 @@
-export * from "./Legend";
